[mem][index] Init Intersect of postings #27
- it is actually called 'adaptive list intersection'
- some naive optimization
  - sort the lists by their length, not just pick the shortest one as
eliminator, but also try to probe lists that are less likely to contain
an element first. (this is naive because the value range may not be
same as list length)
  - store the last probe position, so for next element, the search can
be applied to a smaller range
  - use exponential search instead of binary search for better cache
locality
- there is a paper about it `Improving the Performance of List
Intersection`, most naive optimization are mentioned by it, I am not
sure if I will fully implement the algorithm they suggest, it's a bit
overkill I think. (a.k.a I don't understand it)